sourcecode with linenumbers as table with one row per line: https://github.com/metanorma/metanorma-ogc/issues/465
